+++ __tests__/react.test.js
@@ -1,6 +1,7 @@
 const reactConfig = require('../react');
 const a11yRuleset = require('../rules/react/a11y');
 const baseRuleset = require('../rules/react/base');
+const deprecatedRuleset = require('../rules/deprecated/react');
 const hooksRuleset = require('../rules/react/hooks');
 const jsxRuleset = require('../rules/react/jsx');
 const noPrettierRuleset = require('../rules/react/no-prettier');
@@ -13,13 +14,14 @@ describe('React config', () => {
   it('includes the React rulesets', () => {
     const extension = reactConfig.extends;
 
-    expect(extension).toHaveLength(5);
+    expect(extension).toHaveLength(6);
 
-    expect(extension[0].endsWith('rules/react/a11y.js')).toEqual(true);
-    expect(extension[1].endsWith('rules/react/base.js')).toEqual(true);
-    expect(extension[2].endsWith('rules/react/hooks.js')).toEqual(true);
-    expect(extension[3].endsWith('rules/react/jsx.js')).toEqual(true);
-    expect(extension[4].endsWith('rules/react/no-prettier.js')).toEqual(true);
+    expect(extension[0].endsWith('rules/deprecated/react.js')).toEqual(true);
+    expect(extension[1].endsWith('rules/react/a11y.js')).toEqual(true);
+    expect(extension[2].endsWith('rules/react/base.js')).toEqual(true);
+    expect(extension[3].endsWith('rules/react/hooks.js')).toEqual(true);
+    expect(extension[4].endsWith('rules/react/jsx.js')).toEqual(true);
+    expect(extension[5].endsWith('rules/react/no-prettier.js')).toEqual(true);
   });
 
   it('loads the React ESLint plugin', () => {
@@ -53,6 +55,12 @@ describe('React accessibility ruleset', () => {
 describe('React base ruleset', () => {
   it('loads without error', () => {
     expect(() => baseRuleset).not.toThrow();
+  });
+});
+
+describe('Deprecated React ruleset', () => {
+  it('loads without error', () => {
+    expect(() => deprecatedRuleset).not.toThrow();
   });
 });